format: 검색 기능 하는 Input 컴포넌트 분리
